+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.7.4
+* fixed close button when in fullscreen
+
+## 1.7.2
+* added close button on cupertino & material player
+
 ## 1.7.0
 * 🛠️ [#754](https://github.com/fluttercommunity/chewie/pull/754): Upgraded `wakelock_plus` to version `1.1.0`. Thanks [diegotori](https://github.com/diegotori).
 * **BREAKING CHANGE**: Library now requires at least Dart and Flutter versions `2.18` and `3.3.0` respectively.